utils == helpers (#2809)
